Merging Develop into Release 1.0  From now on anything for release 1.0 will be changed in that branch

+++ include/MINX/Media/Song.h
@@ -28,8 +28,12 @@ namespace MINX
 {
 	namespace Media
 	{
+		/** Represents a song loaded from disk.
+		 */
 		class MINX_API Song
 		{
+			/** Constructs a song from the given file.
+			 */
 			Song(char* fileLocation);
 
 			/* Destroy the sound file safely
@@ -55,15 +59,17 @@ namespace MINX
 			/* Set the Playback volume of the sound file
 			*	@param volume A double value from 0.0 to 100.0 for the GAIN or volume of the sound file
 			*/
-			void SetVolume(double volume);
+			void SetVolume(float volume);
 
 			void SetLoop(bool loop);
 		private:
 
-			/* The volume of the sound file
+			/** The volume of the sound file
 			*/
 			double volume;
-
+			
+			/** The internal sfml object
+			 */
 			sf::Music song;
 		};
 	}
