+++ include/MINX/Media/SoundFile.h
@@ -38,39 +38,43 @@ namespace MINX
 			*	@param fileLocation The path of the file to load.
 			*/
 			SoundFile(char* fileLocation);
-			
+
 			/* Destroy the sound file safely
 			*/
 			~SoundFile();
-			
+
 			/* unload the sound file safely
 			*/
 			void Unload();
-			
+
 			/* Start playback of the sound file
 			*/
 			void Play();
-			
+
 			/* Pause playback of the sound file
 			*/
 			void Pause();
-			
+
 			/* Stop playback of the sound file
 			*/
 			void Stop();
-			
+
 			/* Set the Playback volume of the sound file
 			*	@param volume A double value from 0.0 to 100.0 for the GAIN or volume of the sound file
 			*/
-			void SetVolume(double volume);
+			void SetVolume(float volume);
 		private:
 
 			/* The volume of the sound file
 			*/
 			double volume;
 
+			/** The internal sfml sound buffer
+			 */
 			sf::SoundBuffer soundBuffer;
 
+			/** The internal sfml sound object
+			 */
 			sf::Sound sound;
 		};
 	}
